requirements.txt
Flask==2.0.3
httpx==0.21.1
beautifulsoup4==4.10.0
g4f==0.1.0  # Убедитесь, что это правильная версия, если библиотека доступна
python-telegram-bot==13.7
nest_asyncio==1.5.4